Add iree_ref for gemm functionality test
This PR adds utils to generate an iree reference
mmt kernel that can be compiled and run to
obtain a golden output for comparison. Compared
to the torch output, the advantage of this output
is that it can be matched exactly and makes for
an easier comparison target.
