+++ shark_turbine/kernel/wave/utils.py
@@ -21,6 +21,15 @@ from iree.compiler.dialects.transform import (
 )
 
 import sympy
+import torch
+from iree.compiler import compile_str
+import iree.runtime as rt
+import iree.runtime.benchmark as bench
+
+# TODO: Monkey-patching f16 support, need to fix in iree.
+import numpy
+
+bench.DTYPE_TO_ABI_TYPE[numpy.dtype(numpy.float16)] = "f16"
 
 
 def canonicalize_module(module: Operation):
@@ -181,3 +190,118 @@ def remove_global_indexing(
         for constraint in tilingConstraints:
             new_index[key] = new_index[key].subs({constraint.induction_var: 0})
     return new_index
+
+
+def _invoke(vm_context, device, entry_function, inputs, outputs):
+    arg_list = rt.VmVariantList(len(inputs))
+    ret_list = rt.VmVariantList(len(outputs))
+
+    for input in inputs:
+        input_cpu = input.cpu().contiguous()
+        device_array = rt.asdevicearray(device, input_cpu)
+        arg_list.push_ref(device_array._buffer_view)
+
+    vm_context.invoke(entry_function, arg_list, ret_list)
+
+    for i, ret in enumerate(outputs):
+        device_buffer_view = rt.HalBufferView.__iree_vm_cast__(ret_list.get_as_ref(i))
+        device_array = rt.DeviceArray(device, device_buffer_view)
+
+        # TODO: Make to_host accept out array/buffer, so we can avoid extra data copy.
+        host_array = device_array.to_host()
+
+        # Convert to torch tensor without actually importing torch.
+        ret[:] = type(ret)(host_array)
+
+
+def _write_file(name, mode, data):
+    with open(name, mode) as file:
+        file.write(data)
+
+
+def _print_bench_result(result, filename):
+    import json
+
+    res = json.dumps(result, sort_keys=True, indent=4)
+    if filename is not None:
+        _write_file(filename, "w", res)
+    else:
+        print(res)
+
+
+def compile_and_invoke(
+    asm: str,
+    func_name: str,
+    config: dict[str, str],
+    kernel_inputs: list[torch.Tensor],
+    kernel_outputs: list[torch.Tensor],
+    run: bool = False,
+    run_bench: bool = False,
+):
+    backend = config["backend"]
+    device = config["device"]
+    flags = [
+        f"--iree-hal-target-backends={backend}",
+        "--iree-vm-bytecode-module-strip-source-map=true",
+        "--iree-opt-strip-assertions=true",
+        "--iree-vm-target-truncate-unsupported-floats",
+    ]
+
+    # TODO: More targets/backends support.
+    if backend == "rocm":
+        target = config["target"]
+        flags.append(f"--iree-rocm-target-chip={target}")
+
+    if config.get("print_ir_after_all", False):
+        flags.append("--mlir-print-ir-after-all")
+
+    if run_bench:
+        bench_batch_size = config.get("benchmark_batch_size", None)
+        bench_repetitions = config.get("benchmark_repetitions", None)
+        bench_file = config.get("benchmark_results_file", None)
+
+        benchmark_flags = {}
+
+        if bench_batch_size is not None:
+            flags.append(
+                f"--iree-hal-benchmark-dispatch-repeat-count={bench_batch_size}"
+            )
+            benchmark_flags["batch_size"] = int(bench_batch_size)
+
+        if bench_repetitions is not None:
+            benchmark_flags["benchmark_repetitions"] = int(bench_repetitions)
+
+    res = compile_str(asm, target_backends=[backend], extra_args=flags)
+
+    dump_vmfb_file = config.get("dump_vmfb_file", None)
+    if dump_vmfb_file is not None:
+        _write_file(dump_vmfb_file, "wb", res)
+
+    rt_config = rt.Config(device)
+    device = rt_config.device
+    vm_instance = rt_config.vm_instance
+    mod = rt.VmModule.copy_buffer(vm_instance, res)
+
+    vm_modules = [
+        mod,
+        rt.create_hal_module(vm_instance, device),
+    ]
+    ctx = rt.SystemContext(
+        vm_modules=vm_modules,
+        config=rt_config,
+    )
+
+    if run:
+        func = mod.lookup_function(func_name)
+        _invoke(ctx.vm_context, device, func, kernel_inputs, kernel_outputs)
+
+    if run_bench:
+        inputs = [inp.numpy() for inp in kernel_inputs]
+        benchmark_results = bench.benchmark_module(
+            mod,
+            entry_function=func_name,
+            device=device,
+            inputs=inputs,
+            **benchmark_flags,
+        )
+        _print_bench_result(benchmark_results, bench_file)
